+++ app/controllers/backoffice/application_controller.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+class Backoffice::ApplicationController < ApplicationController
+  before_action :authenticate_user!
+  before_action :backoffice_authorization!
+
+  layout "backoffice"
+
+  def policy_scope(scope)
+    super([:backoffice, scope])
+  end
+
+  def authorize(record, query = nil)
+    super([:backoffice, record], query)
+  end
+
+  def permitted_attributes(record, action = action_name)
+    super([:backoffice, record], action)
+  end
+
+  private
+
+    def backoffice_authorization!
+      authorize :backoffice, :show?
+    end
+end