+++ app/policies/backoffice/service_policy.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+class Backoffice::ServicePolicy < ApplicationPolicy
+  class Scope < Scope
+    def resolve
+      scope.where(owner: user)
+    end
+  end
+
+  def index?
+    true
+  end
+
+  def show?
+    owner?
+  end
+
+  def new?
+    true
+  end
+
+  def create?
+    true
+  end
+
+  def update?
+    owner?
+  end
+
+  def destroy?
+    owner? && record.orders.count.zero?
+  end
+
+  def permitted_attributes
+    [:title, :description, :terms_of_use]
+  end
+
+  private
+
+    def owner?
+      record.owner == user
+    end
+end