+++ app/services/service/create.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Service::Create
+  def initialize(service)
+    @service = service
+  end
+
+  def call
+    @service.save
+
+    @service
+  end
+end