+++ db/migrate/20180706090424_add_owner_to_service.rb
@@ -1,0 +1,6 @@
+class AddOwnerToService < ActiveRecord::Migration[5.2]
+  def change
+    add_reference :services, :owner
+    add_foreign_key :services, :users, column: :owner_id
+  end
+end