+++ db/migrate/20180706101250_add_roles_mask_to_user.rb
@@ -1,0 +1,5 @@
+class AddRolesMaskToUser < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :roles_mask, :integer
+  end
+end