+++ spec/models/service_spec.rb
@@ -7,6 +7,8 @@ RSpec.describe Service do
   it { should validate_presence_of(:description) }
   it { should validate_presence_of(:tagline) }
 
+  it { should belong_to(:owner) }
+
   it { should have_many(:service_categories).dependent(:destroy) }
   it { should have_many(:categories) }
   it { should have_many(:orders) }